% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ClusterCount.R
\name{ClusterCount}
\alias{ClusterCount}
\title{ClusterCount}
\usage{
ClusterCount(compounds, cpd.identifier, height = 0.8)
}
\arguments{
\item{compounds}{data table of compounds from a specific plate}

\item{cpd.identifier}{character string identifying compounds: Broad ID, InChIKey, etc...}

\item{height}{numeric value to cut dendrogram defaults to 0.8}
}
\description{
Counts the number of clusters at a specfied height on a dendrogram
generated from an ECFP distance matrix
}
\examples{
hashes <- round(rnorm(50, 100000000, 1000), 1)

compounds <- data.table('ECFP_4[1]' = sample(hashes, 100, replace=TRUE),
                        'ECFP_4[2]' = sample(hashes, 100, replace=TRUE),
                        'ECFP_4[3]' = sample(hashes, 100, replace=TRUE),
                        'ECFP_4[4]' = sample(hashes, 100, replace=TRUE),
                        'ECFP_4[5]' = sample(hashes, 100, replace=TRUE))

compounds$InChIKey <- sapply(1:dim(compounds)[1], function(x)
                             paste(sample(LETTERS, 10, replace=TRUE), collapse=''))

ClusterCount(compounds, 'InChIKey')
}

