% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CrossDiverse.R
\name{CrossDiverse}
\alias{CrossDiverse}
\title{CrossDiverse}
\usage{
CrossDiverse(candidate, selection, cpd.identifier, height = 0.5)
}
\arguments{
\item{candidate}{data frame or data table of candidate compounds with ECFP finger prints and compound identifier columns}

\item{height}{numeric value to cut dendrogram defaults to 0.5}

\item{reference}{data frame or data table of candidate compounds with ECFP finger prints and compound identifier columns}
}
\description{
Find candidate compounds unique to a reference set
}
\examples{
hashes <- round(rnorm(25, 100000000, 1000), 1)

plate.a <- data.table('ECFP_4[1]' = sample(hashes, 100, replace=TRUE),
                      'ECFP_4[2]' = sample(hashes, 100, replace=TRUE),
                      'ECFP_4[3]' = sample(hashes, 100, replace=TRUE),
                      'ECFP_4[4]' = sample(hashes, 100, replace=TRUE),
                      'ECFP_4[5]' = sample(hashes, 100, replace=TRUE))

plate.a$InChIKey <- sapply(1:dim(plate.a)[1], function(x)
                           paste(sample(LETTERS, 10, replace=TRUE), collapse=''))

plate.b <- data.table('ECFP_4[1]' = sample(hashes, 100, replace=TRUE),
                      'ECFP_4[2]' = sample(hashes, 100, replace=TRUE),
                      'ECFP_4[3]' = sample(hashes, 100, replace=TRUE),
                      'ECFP_4[4]' = sample(hashes, 100, replace=TRUE),
                      'ECFP_4[5]' = sample(hashes, 100, replace=TRUE))

plate.b$InChIKey <- sapply(1:dim(plate.b)[1], function(x)
                           paste(sample(LETTERS, 10, replace=TRUE), collapse=''))

CrossDiverse(plate.a, plate.b, cpd.identifier='InChIKey')
}

