% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sfp.R
\name{sfp}
\alias{sfp}
\title{sfp}
\usage{
sfp(data, cpd.identifier)
}
\arguments{
\item{data}{data.table containing compound identifier and fingerprint hash}

\item{cpd.identifier}{character string identifying compounds: Broad ID, InChIKey, etc...}
}
\description{
Converts ECFP hash codes into a sparse matrix by arbitrarily assigning them as indices.
}
\examples{
hashes <- round(rnorm(120, 100000000, 1000), 1)
compounds <- data.table('ECFP_4[1]' = sample(hashes, 100, replace=TRUE),
                        'ECFP_4[2]' = sample(hashes, 100, replace=TRUE),
                        'ECFP_4[3]' = sample(hashes, 100, replace=TRUE),
                        'ECFP_4[4]' = sample(hashes, 100, replace=TRUE),
                        'ECFP_4[5]' = sample(hashes, 100, replace=TRUE))
compounds$InChIKey <- sapply(1:dim(compounds)[1], function(x)
                             paste(sample(LETTERS, 10, replace=TRUE), collapse=''))
sparseFP <- sfp(compounds, cpd.identifier='InChIKey')
sparseFP[1:10, 1:15]
}

